Updating release to allow maven to publish (#95)
When manually triggering a workflow, the publish to maven step is always skipped despite having the value checked. I believe this is due to `github.events.inputs` context converting the value to a string and the check itself is comparing against a bool. Github documentation states: 

```
Note: The workflow will also receive the inputs in the github.event.inputs context. The information in the inputs context and github.event.inputs context is identical except that the inputs context preserves Boolean values as Booleans instead of converting them to strings.
```

So changing contexts should let us check the bool directly.